Extract cores data

+++ src/PcStatsReporter.Core/Models/CpuCore.cs
@@ -7,7 +7,7 @@ namespace PcStatsReporter.Core.Models
         public uint Id { get; set; }
         public uint Temperature { get; set; }
         public uint Speed { get; set; }
-        public double Load { get; set; }
+        public uint Load { get; set; }
 
         public override string ToString()
         {
